requirements.txt
torch==2.0.1
transformers==4.33.0
Pillow==10.0.0
pytest==7.4.0